Replace `MaybeUninit` with `Option` to remove most unsafe blocks in `transformers.rs`
I see little reasons to use `MaybeUninit` here:
- `Option<&T>` has exactly the same size as `&T`, because references cannot be null (same as [num::NonZero](https://doc.rust-lang.org/std/num/struct.NonZero.html).
- using `expect` instead of `assume_init` adds an additional check, which may be ellided or not by the compiler, but in practice even if the check remains the branch predictor is always going to take the non-failing branch by default so the slowdown should be minimal.

Anecdotally, this change had no impact on the token throughput on my machine.

If the supposedly very low performance cost was deemed to high anyway, the correct solution would be to “make invalid state unrepresentable” by refactoring the code in order to get rid of the Option, but that's a bit more work and that would change the structure of the code quite a bit, which I didn't want to do.